(#13) Add Elastic APM Go agent to get traces from Gin-Gonic
Package apmgin provides middleware for the Gin web framework.

For each request, a transaction is stored in the request context,
which can be obtained via gin.Context.Request.Context() in your handler.

The apmgin middleware will recover panics and send them to Elastic APM,
so you do not need to install the gin.Recovery middleware.