(#13) Add APM Elasticsearch integration
It provides a means of instrumenting the HTTP transport of Elasticsearch
clients, so that Elasticsearch requests are reported as spans within the
current transaction.

To create spans for an Elasticsearch request, you should wrap the client’s
HTTP transport using the WrapRoundTripper function, and then associate the
request with a context containing a transaction.